% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nupputils.R
\name{build_pages}
\alias{build_pages}
\title{Construir paginas}
\usage{
build_pages(files, cores = 10)
}
\arguments{
\item{files}{Arquivos}

\item{cores}{Nucleos de processamento}
}
\description{
Construir paginas
}
