% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfal.R
\name{download_and_parse_jfal}
\alias{download_and_parse_jfal}
\title{Baixar e parsear processos da JFAL}
\usage{
download_and_parse_jfal(id, path_rds = ".")
}
\arguments{
\item{id}{Numero do processo}

\item{path_rds}{Caminho onde salvar os RDSs}
}
\description{
Diferentemente da maioria dos downloaders/parsers,
a organizacao da JFAL faz com que esses passos precisem estar juntos.
Aqui o objeto salvo ja representa o processo parseado.
}
\examples{
\dontrun{
download_and_parse_jfal("0009214-22.2005.4.05.8000")
}

}
