% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfrj.R
\name{download_and_parse_jfrj}
\alias{download_and_parse_jfrj}
\title{Baixar e parsear processos da JFRJ}
\usage{
download_and_parse_jfrj(id, path_rds = ".")
}
\arguments{
\item{id}{Numero do processo}

\item{path_rds}{Caminho onde salvar os RDSs}
}
\description{
Diferentemente da maioria dos downloaders/parsers,
a organizacao da JFRJ faz com que esses passos precisem estar juntos.
Aqui o objeto salvo ja representa o processo parseado.
}
\examples{
\dontrun{
download_and_parse_jfrj("01885848620174025101")
}

}
