% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfsp.R
\name{download_and_parse_jfsp}
\alias{download_and_parse_jfsp}
\title{Baixar e parsear processos da JFSP}
\usage{
download_and_parse_jfsp(id, path_rds = ".", detalhe_fases = FALSE)
}
\arguments{
\item{id}{Numero do processo}

\item{path_rds}{Caminho onde salvar os RDSs}

\item{detalhe_fases}{Se deve ou nao extrair os detalhes das fases}
}
\description{
Diferentemente da maioria dos downloaders/parsers,
a organizacao da JFSP faz com que esses passos precisem estar juntos.
Aqui o objeto salvo ja representa o processo parseado.
}
\examples{
\dontrun{
download_and_parse_jfsp("0001274-57.1993.4.03.6100")
}

}
