% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf2.R
\name{download_and_parse_trf2}
\alias{download_and_parse_trf2}
\title{Baixar e parsear processos do TRF2}
\usage{
download_and_parse_trf2(id, path_rds = ".")
}
\arguments{
\item{id}{Numero do processo}

\item{path_rds}{Caminho onde salvar os RDSs}
}
\description{
Diferentemente da maioria dos downloaders/parsers,
a organizacao do TRF2 faz com que esses passos precisem estar juntos.
Aqui o objeto salvo ja representa o processo parseado.
}
\examples{
\dontrun{
download_and_parse_trf2("0011542-27.2017.4.02.0000")
}

}
