% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfdf.R
\name{download_jfdf}
\alias{download_jfdf}
\title{Baixar processo da JFDF}
\usage{
download_jfdf(id, path = ".")
}
\arguments{
\item{id}{Numero do processo}

\item{path}{Caminho onde salvar HTMLs}
}
\description{
Baixar processo da JFDF
}
\examples{
\dontrun{
download_jfdf("0036535-45.2000.4.01.3400")
}

}
