% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjal.R
\name{download_tjal}
\alias{download_tjal}
\title{Baixar processos do TJAL}
\usage{
download_tjal(id, path = ".", inst = "1")
}
\arguments{
\item{id}{Numero do processo}

\item{path}{Caminho para a pasta onde salvar os HTMLs}

\item{inst}{Numero da instancia (1 ou 2)}
}
\description{
Baixar processos do TJAL
}
\examples{
\dontrun{
download_tjal("0803813-44.2018.8.02.0000", inst = 2)
}

}
