% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjdft.R
\name{download_tjdft}
\alias{download_tjdft}
\title{Baixar processos do TJDFT}
\usage{
download_tjdft(id, path = ".", inst = "1")
}
\arguments{
\item{id}{Numero do processo}

\item{path}{Caminho para a pasta onde salvar os HTMLs}

\item{inst}{Numero da instancia (1 ou 2)}
}
\description{
Baixar processos do TJDFT
}
\examples{
\dontrun{
download_tjdft("0003100-78.2016.8.07.0018")
}

}
