% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjrj.R
\name{download_tjrj}
\alias{download_tjrj}
\title{Baixar processos do TJRJ}
\usage{
download_tjrj(id, path = ".")
}
\arguments{
\item{id}{Numero do processo}

\item{path}{Caminho para a pasta onde salvar os HTMLs}
}
\description{
Baixar processos do TJRJ
}
\examples{
\dontrun{
download_tjrj("00427318920128190000")
}

}
