% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf1.R
\name{download_trf1}
\alias{download_trf1}
\title{Baixar processos do TRF1}
\usage{
download_trf1(id, path = ".")
}
\arguments{
\item{id}{Numero do processo}

\item{path}{Caminho para a pasta onde salvar os HTMLs}
}
\description{
Baixar processos do TRF1
}
\examples{
\dontrun{
download_trf1("0009384-30.2011.4.01.3300")
}

}
