% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf5.R
\name{download_trf5}
\alias{download_trf5}
\title{Baixar processos do TRF5}
\usage{
download_trf5(id, path = ".")
}
\arguments{
\item{id}{Numero do processo}

\item{path}{Caminho para a pasta onde salvar os HTMLs}
}
\description{
Baixar processos do TRF5
}
\examples{
\dontrun{
download_trf5("0000193-94.2016.4.05.8401")
}

}
