% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjal.R
\name{parse_tjal}
\alias{parse_tjal}
\title{Parsear processos do TJAL}
\usage{
parse_tjal(file, inst = 1)
}
\arguments{
\item{file}{Caminho para o processo}

\item{inst}{Numero da instancia (1 ou 2)}
}
\description{
Parsear processos do TJAL
}
