% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjdft.R
\name{parse_tjdft}
\alias{parse_tjdft}
\title{Parsear processos do TJDFT}
\usage{
parse_tjdft(file, inst = "1")
}
\arguments{
\item{file}{Caminho para o processo}

\item{inst}{Numero da instancia (1 ou 2)}
}
\description{
Parsear processos do TJDFT
}
