% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trf1.R
\name{parse_trf1}
\alias{parse_trf1}
\title{Parsear processos do TRF1}
\usage{
parse_trf1(file_movs)
}
\arguments{
\item{file_movs}{Caminho para os movimentos do processo}
}
\description{
Parsear processos do TRF1
}
