% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nupputils.R
\name{pesquisar_assunto}
\alias{pesquisar_assunto}
\title{Pesquisar assunto}
\usage{
pesquisar_assunto(x, txt_data = all_texts_pags_clean)
}
\arguments{
\item{x}{Assunto}

\item{txt_data}{Dados do texto}
}
\description{
Pesquisar assunto
}
