% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nupputils.R
\name{read_all_txt_parallel}
\alias{read_all_txt_parallel}
\title{Ler textos em paralelo}
\usage{
read_all_txt_parallel(all_text_files)
}
\arguments{
\item{all_text_files}{Todos os arquivos}
}
\description{
Ler textos em paralelo
}
