% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nupputils.R
\name{write_pages}
\alias{write_pages}
\title{Escrever paginas}
\usage{
write_pages(arqs_list, chunks = 100)
}
\arguments{
\item{arqs_list}{Lista de arquivos}

\item{chunks}{Numero de blocos}
}
\description{
Escrever paginas
}
